requirements.txt
django >=5.0.0, < 5.1.1
djangorestframework
pyyaml
requests
django-cors-headers
django-injector
inject
pandas
dateparser
python-dotenv
mysqlclient
pylibmagic
python-magic